requirements.txt
requests==2.32.5
polars==1.36.1
fsspec==2026.2.0
adlfs==2026.2.0
azure-storage-blob==12.28.0
python-dotenv==1.2.2
pandas==3.0.1